+++ debug.txt
@@ -1,0 +1,18 @@
+For debugging child process:
+
+>>>>>>
+
+{
+	// https://sourceware.org/gdb/onlinedocs/gdb/Forks.html
+	"description": "Fork follows Child process",
+	"text": "set follow-fork-mode child",
+	"ignoreFailures": true
+},
+{
+	// https://sourceware.org/gdb/onlinedocs/gdb/Forks.html
+	"description": "Fork will keep the other process attached to debugger",
+	"text": "set detach-on-fork off",
+	"ignoreFailures": true
+}
+
+<<<<<<
