+++ src/execution/_child_execute_target_command.c
@@ -1,37 +1,31 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   execute_target_command.c                           :+:      :+:    :+:   */
+/*   _child_execute_target_command.c                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: Philip <juli@student.42london.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/17 15:59:30 by Philip            #+#    #+#             */
-/*   Updated: 2024/04/19 21:11:54 by Philip           ###   ########.fr       */
+/*   Updated: 2024/04/20 00:38:11 by Philip           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../command_list/cmd_list.h"
+#include "../minishell/minishell.h" /* minishell */
+#include "../command_list/t_cmd_list.h" /* t_cmd_list */
 #include "../environment_variables/env.h" /* env_build_envp */
-#include "../minishell/minishell.h"
-#include "../exit_status.h"
-#include "../free/free.h" /* free_cmds_env_pipes_rl_clear_history */
+#include "../exit_status.h" /* exit_status */
+#include "../pipes/t_pipes.h" /* t_pipes */
+#include "../free/free.h" /*  */
 #include "file_check.h" /* file_check */
 #include "libft.h" /* ft_dprintf */
 #include <unistd.h> /* access, F_OK, STDERR_FILENO, execve */
 #include <stdlib.h> /* exit */
 #include <readline/readline.h> /* rl_clear_history */
 
-int	_command_not_found(const char *command)
-{
-	ft_dprintf(STDERR_FILENO, "minishell: %s: command not found\n", command);
-	return (COMMAND_NOT_FOUND_EXIT_STATUS);
-}
-
-int	_is_a_directory(const char *command)
-{
-	ft_dprintf(STDERR_FILENO, "minishell: %s: Is a directory\n", command);
-	return (IS_A_DIRECTORY_EXIT_STATUS);
-}
+extern void	_child_execute_target_command(t_cmd_list *cmd, t_env *env,
+				t_pipes *pipes);
+static int	_command_not_found(const char *command);
+static int	_is_a_directory(const char *command);
 
 /**
  * @brief Executes the target command. This function exits the process and frees
@@ -66,4 +60,16 @@ extern void	_child_execute_target_command(t_cmd_list *cmd, t_env *env,
 		free(pipes->pipes);
 	minishell_free();
 	exit(exit_status);
+}
+
+static int	_command_not_found(const char *command)
+{
+	ft_dprintf(STDERR_FILENO, "minishell: %s: command not found\n", command);
+	return (COMMAND_NOT_FOUND_EXIT_STATUS);
+}
+
+static int	_is_a_directory(const char *command)
+{
+	ft_dprintf(STDERR_FILENO, "minishell: %s: Is a directory\n", command);
+	return (IS_A_DIRECTORY_EXIT_STATUS);
 }
