+++ src/functions.h
@@ -6,7 +6,7 @@
 /*   By: Philip <juli@student.42london.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/03 20:43:44 by Philip            #+#    #+#             */
-/*   Updated: 2024/04/16 14:20:17 by Philip           ###   ########.fr       */
+/*   Updated: 2024/04/19 22:15:07 by Philip           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -42,7 +42,7 @@ void		check_redirect_files(t_cmd_list *cmds);
 
 /* Execute commands */
 
-void		execute_cmds(t_cmd_list *cmds, t_env **env);
+void		execute_commands(t_cmd_list *cmds, t_env **env);
 
 /* Test functions */
